requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
weasyprint==60.2
python-pptx==0.6.23
httpx==0.26.0
pytest==7.4.4
requests==2.31.0

